requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.3.3
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
prometheus-client==0.20.0
psutil==6.0.0
opencv-python==4.10.0.84
ultralytics==8.3.20
Pillow==10.4.0
httpx==0.27.0
numpy==1.26.4
